\name{multicore.process}
\Rdversion{1.1}
\docType{class}
\alias{multicore.process-class}
\alias{multicore.process}

\title{Class ``multicore.process''}
\description{The \code{multicore.process} class provides a reference object
implementation of \code{process.interface} class.  It uses the multicore
package to background calculations. 
}
\keyword{classes}
